Rollback the UniqueAtomFilter changes - it makes sense to keep behaviour aligned with convention of other toolkits. Added a test to avoid making the change in future.
